+++ src/limbo/opt/grid_search.hpp
@@ -8,11 +8,11 @@
 #include <limbo/tools/macros.hpp>
 
 namespace limbo {
-   namespace defaults {
+    namespace defaults {
         struct opt_gridsearch {
             BO_PARAM(int, bins, 5);
         };
-    }    
+    }
     namespace opt {
         template <typename Params>
         struct GridSearch {
@@ -27,7 +27,7 @@ namespace limbo {
 
         protected:
             template <typename F>
-            Eigen::VectorXd _inner_search(const F& f, int depth, const Eigen::VectorXd& current) const
+            Eigen::VectorXd _inner_search(const F& f, size_t depth, const Eigen::VectorXd& current) const
             {
                 double step_size = 1.0 / (double)Params::opt_gridsearch::bins();
                 double upper_lim = 1.0 + step_size;
